+++ lib/ransack/adapters/active_record/context.rb
@@ -97,8 +97,7 @@ module Ransack
         # JoinDependency to track table aliases.
         #
         def join_sources
-          base, joins =
-          if ::ActiveRecord::VERSION::STRING > Constants::RAILS_5_2_0
+          base, joins = begin
             alias_tracker = ::ActiveRecord::Associations::AliasTracker.create(self.klass.connection, @object.table.name, [])
             constraints   = if ::Gem::Version.new(::ActiveRecord::VERSION::STRING) >= ::Gem::Version.new(Constants::RAILS_6_0)
               @join_dependency.join_constraints(@object.joins_values, alias_tracker)
@@ -109,11 +108,6 @@ module Ransack
             [
               Arel::SelectManager.new(@object.table),
               constraints
-            ]
-          else
-            [
-              Arel::SelectManager.new(@object.table),
-              @join_dependency.join_constraints(@object.joins_values, @join_type)
             ]
           end
           joins.each do |aliased_join|
@@ -260,23 +254,15 @@ module Ransack
 
           join_list = join_nodes + convert_join_strings_to_ast(relation.table, string_joins)
 
-          if ::ActiveRecord::VERSION::STRING == Constants::RAILS_5_2_0
-            alias_tracker = ::ActiveRecord::Associations::AliasTracker.create(self.klass.connection, relation.table.name, join_list)
-            join_dependency = Polyamorous::JoinDependency.new(relation.klass, relation.table, association_joins, alias_tracker)
-            join_nodes.each do |join|
-              join_dependency.send(:alias_tracker).aliases[join.left.name.downcase] = 1
-            end
+          alias_tracker = ::ActiveRecord::Associations::AliasTracker.create(self.klass.connection, relation.table.name, join_list)
+          join_dependency = if ::Gem::Version.new(::ActiveRecord::VERSION::STRING) >= ::Gem::Version.new(Constants::RAILS_6_0)
+            Polyamorous::JoinDependency.new(relation.klass, relation.table, association_joins, Arel::Nodes::OuterJoin)
           else
-            alias_tracker = ::ActiveRecord::Associations::AliasTracker.create(self.klass.connection, relation.table.name, join_list)
-            join_dependency = if ::Gem::Version.new(::ActiveRecord::VERSION::STRING) >= ::Gem::Version.new(Constants::RAILS_6_0)
-              Polyamorous::JoinDependency.new(relation.klass, relation.table, association_joins, Arel::Nodes::OuterJoin)
-            else
-              Polyamorous::JoinDependency.new(relation.klass, relation.table, association_joins)
-            end
-            join_dependency.instance_variable_set(:@alias_tracker, alias_tracker)
-            join_nodes.each do |join|
-              join_dependency.send(:alias_tracker).aliases[join.left.name.downcase] = 1
-            end
+            Polyamorous::JoinDependency.new(relation.klass, relation.table, association_joins)
+          end
+          join_dependency.instance_variable_set(:@alias_tracker, alias_tracker)
+          join_nodes.each do |join|
+            join_dependency.send(:alias_tracker).aliases[join.left.name.downcase] = 1
           end
           join_dependency
         end
@@ -308,15 +294,6 @@ module Ransack
               @join_type
             )
             found_association = jd.instance_variable_get(:@join_root).children.last
-          elsif ::ActiveRecord::VERSION::STRING == Constants::RAILS_5_2_0
-            alias_tracker = ::ActiveRecord::Associations::AliasTracker.create(self.klass.connection, parent.table.name, [])
-            jd = Polyamorous::JoinDependency.new(
-              parent.base_klass,
-              parent.table,
-              Polyamorous::Join.new(name, @join_type, klass),
-              alias_tracker
-            )
-            found_association = jd.instance_variable_get(:@join_root).children.last
           else
             jd = Polyamorous::JoinDependency.new(
               parent.base_klass,
@@ -333,11 +310,7 @@ module Ransack
           @join_dependency.instance_variable_get(:@join_root).children.push found_association
 
           # Builds the arel nodes properly for this association
-          if ::ActiveRecord::VERSION::STRING > Constants::RAILS_5_2_0
-            @join_dependency.send(:construct_tables!, jd.instance_variable_get(:@join_root))
-          else
-            @join_dependency.send(:construct_tables!, jd.instance_variable_get(:@join_root), found_association)
-          end
+          @join_dependency.send(:construct_tables!, jd.instance_variable_get(:@join_root))
 
           # Leverage the stashed association functionality in AR
           @object = @object.joins(jd)
@@ -347,22 +320,12 @@ module Ransack
         def extract_joins(association)
           parent = @join_dependency.instance_variable_get(:@join_root)
           reflection = association.reflection
-          join_constraints = if ::ActiveRecord::VERSION::STRING <= Constants::RAILS_5_2_0
-                               association.join_constraints(
-                                 parent.table,
-                                 parent.base_klass,
-                                 Arel::Nodes::OuterJoin,
-                                 association.tables,
-                                 reflection.chain
-                               )
-                             else
-                               association.join_constraints(
+          join_constraints = association.join_constraints(
                                  parent.table,
                                  parent.base_klass,
                                  Arel::Nodes::OuterJoin,
                                  @join_dependency.instance_variable_get(:@alias_tracker)
                                )
-                             end
           join_constraints.to_a.flatten
         end
       end