+++ lib/ransack/constants.rb
@@ -45,8 +45,6 @@ module Ransack
     NOT_EQ_ALL          = 'not_eq_all'.freeze
     CONT                = 'cont'.freeze
 
-    RAILS_5_2           = '5.2'.freeze
-    RAILS_5_2_0         = '5.2.0'.freeze
     RAILS_6_0           = '6.0.0'.freeze
 
     RANSACK_SLASH_SEARCHES = 'ransack/searches'.freeze