+++ spec/helpers/polyamorous_helper.rb
@@ -7,18 +7,9 @@ module PolyamorousHelper
     def new_join_dependency(klass, associations = {})
       Polyamorous::JoinDependency.new klass, klass.arel_table, associations, Polyamorous::InnerJoin
     end
-  elsif ActiveRecord.version > ::Gem::Version.new("5.2.0")
-    def new_join_dependency(klass, associations = {})
-      Polyamorous::JoinDependency.new klass, klass.arel_table, associations
-    end
-  elsif ActiveRecord.version == ::Gem::Version.new("5.2.0")
-    def new_join_dependency(klass, associations = {})
-      alias_tracker = ::ActiveRecord::Associations::AliasTracker.create(klass.connection, klass.table_name, [])
-      Polyamorous::JoinDependency.new klass, klass.arel_table, associations, alias_tracker
-    end
   else
     def new_join_dependency(klass, associations = {})
-      Polyamorous::JoinDependency.new klass, associations, []
+      Polyamorous::JoinDependency.new klass, klass.arel_table, associations
     end
   end